Add net embedding logic
When building a net, the returned value is an "embedding" function that
may be used as a composable unit in other nets.

Example:

```
(def n1 (output :n1/out (node (input :n1/in) identity)))
(def n2
  (let [init (input :init)
        a (node init identity)
        {b :n1/out} (n1 {:n1/in a})]
    (output :stdout b))
```

Also, the 2 argument form of node: `(node xf)` will return an embedding
function with a single input `:in` and a single output `:out`. This will
be useful for specifying transducer nets in JSON.

Last: extended the concept of merging to be a tree of sets:

```
;; All nodes have the same input
(node a identity)
(node #{a} identity)
(node #{a #{a}} identity)
```