Add a new, transducer based API to specify nets
This is surely the last API right? The nice thing about this one is
that it allows intermediate inspection of a net since each defined
input/node/output is a transducer that will make a net-map and run it
through netduce.

Also, add filter to match-tags to ignore poorly formatted input

Also, add guards to ensure inputs are present and that those inputs are
net transducers.